test: split PDF list and heading flow cases
